float で計算していた部分を double に修正。これで IPv4 のアドレス範囲ならば大丈夫になるはず。

+++ trunk/PackterViewer/PackterViewer/FlyPacket.cs
@@ -13,22 +13,22 @@ namespace Packter_viewer
 {
     public interface FlyPacket
     {
-        float SrcAddress
+        double SrcAddress
         {
             get;
         }
 
-        float DstAddress
+        double DstAddress
         {
             get;
         }
 
-        float SrcPort
+        double SrcPort
         {
             get;
         }
 
-        float DstPort
+        double DstPort
         {
             get;
         }
